# metal : handle -inf values in FA kernel