@slaren 2024-05-21
The original issue may have been caused by using fast math.

![image](https://github.com/ggerganov/llama.cpp/assets/2141330/fa6b8781-ab14-4d24-8d0d-2ca21110c01f)