@ggerganov 2024-05-21
I tried disabling it, but it still produced NaNs